Build contrail-nodemgr package, enhance packages.make for parallel build
But contrail-database package is not built, so nodemgr for it is not
delivered.

Package contrail-database only provides supervisord config to run cassandra
under supervisor. I think cassandra should be executed by init system and not
by supervisor, unfortunately I am not sure if absence of contrail-database
nodemgr will not be an issue.

I am going to test these packages in real deployment, so we will see..

I would appreciate feedback or someone to discuss contrail-packages building
so please feel free to contact me.